pyproject.toml
[tool.poetry]
name = "iic2343"
version = "0.1.0"
description = "Write to the Basys3 directly."
license = "MIT"
authors = ["Daniel Leal <dlleal@uc.cl>"]
maintainers = ["Daniel Leal <dlleal@uc.cl>"]
readme = "README.md"
exclude = [
    ".github",
    ".flake8",
    ".pylintrc",
    "mypy.ini",
    ".coveragerc",
    "Makefile",
    "scripts",
    "tests"
]

[tool.poetry.dependencies]
python = "^3.6"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.6.4"
mypy = "^0.790"
pylint = "^2.6.0"
pytest = "^6.1.1"
pytest-cov = "^2.10.1"
